Display email verification warning
