+++ web/src/components/Repetition/index.tsx
@@ -16,35 +16,13 @@
  * along with Dnote.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import Helmet from 'react-helmet';
-import { Link } from 'react-router-dom';
-import classnames from 'classnames';
 
-import { getNewRepetitionPath } from 'web/libs/paths';
-import { getRepetitionRules } from '../../store/repetitionRules';
 import PayWall from '../Common/PayWall';
-import { useDispatch, useSelector } from '../../store';
-import RepetitionList from './RepetitionList';
-import DeleteRepetitionRuleModal from './DeleteRepetitionRuleModal';
-import Flash from '../Common/Flash';
-import styles from './Repetition.scss';
+import Content from './Content';
 
 const Repetition: React.FunctionComponent = () => {
-  const dispatch = useDispatch();
-  useEffect(() => {
-    dispatch(getRepetitionRules());
-  }, [dispatch]);
-
-  const { repetitionRules } = useSelector(state => {
-    return {
-      repetitionRules: state.repetitionRules
-    };
-  });
-
-  const [ruleUUIDToDelete, setRuleUUIDToDelete] = useState('');
-  const [successMessage, setSuccessMessage] = useState('');
-
   return (
     <div className="page page-mobile-full">
       <Helmet>
@@ -52,47 +30,7 @@ const Repetition: React.FunctionComponent = () => {
       </Helmet>
 
       <PayWall>
-        <div className="container mobile-fw">
-          <div className={classnames('page-header', styles.header)}>
-            <h1 className="page-heading">Repetition</h1>
-
-            <Link
-              id="T-new-rule-btn"
-              className="button button-first button-normal"
-              to={getNewRepetitionPath()}
-            >
-              New
-            </Link>
-          </div>
-        </div>
-
-        <div className="container mobile-nopadding">
-          <Flash
-            when={successMessage !== ''}
-            kind="success"
-            onDismiss={() => {
-              setSuccessMessage('');
-            }}
-          >
-            {successMessage}
-          </Flash>
-
-          <RepetitionList
-            isFetching={repetitionRules.isFetching}
-            isFetched={repetitionRules.isFetched}
-            items={repetitionRules.data}
-            setRuleUUIDToDelete={setRuleUUIDToDelete}
-          />
-        </div>
-
-        <DeleteRepetitionRuleModal
-          repetitionRuleUUID={ruleUUIDToDelete}
-          isOpen={ruleUUIDToDelete !== ''}
-          onDismiss={() => {
-            setRuleUUIDToDelete('');
-          }}
-          setSuccessMessage={setSuccessMessage}
-        />
+        <Content />
       </PayWall>
     </div>
   );
